initial v1.0.0.0rc commit
The protocol has been renamed from Cardano-Secondary-Market to
Cardano-Aftermarket. Instead of renaming the github repo, the old
[cardano-secondary-market](https://github.com/fallen-icarus/cardano-secondary-market)
repo has been archived and a new repo has been created. This was done to
minimize the risk of repojacking.

This commit includes new features and optimizations, as well as updated
documentation to go with the changes. *All* of the features discussed in
the *Future Directions* section of the previous verion's README have
been added. To see a detailed summary of everything that has changed,
see the CHANGELOG.