NEW FEATURE: Add Claim Auction Type (#1)
The prior version only supports auctions where the seller finalizes the auction and submits the final transaction. This is an issue for certain DApp compositions where the buyer may need to submit the final transaction. An example of this is when a user buys a loan bond for cardano-loans; the user must update the lender payment address for the loan in the same transaction. This is not possible when the seller finalizes the trade.

To enable support for this, *another* auction type was added. The original auction type is now called a *Spot Auction*, and a new *Claim Auction* is one with the following flow:
1. The seller initiates an auction by creating an `Auction` UTxO. This UTxO does *not* specify the type of auction. The same `Auction` UTxO is used for both Spot Auctions and Claim Auctions.
2. The bidder creates a `ClaimBid` UTxO for NFTs at auction. This UTxO includes a deposit (at least the minUTxOValue amount but can be more). It also specifies a time by which the bidder will claim the NFTs, if the bid is accepted.
3. The seller accepts the `ClaimBid` by locking the NFTs in an `AcceptedBid` UTxO for the bidder. The bidder's deposit is locked in this `AcceptedBid` UTxO, as well. The bidder has until their specified "claim expiration" to claim the NFTs.
4. Only that bidder can claim the `AcceptedBid` UTxO and only if they pay the seller the agreed upon bid amount.
5. If the bidder does not claim the NFTs within the agreed upon amount of time, the seller can reclaim the NFTs and confiscate the bidder's deposit.

Bidders decide which bid type to use. A given auction can use both types; the seller can see which type the bid is and can factor this in to their decision of which bid to choose.